requirements.txt
fastapi[standard]
fastapi-cache2[memcache]
uvicorn
schedule
orjson

python-dotenv

discord-webhook
aiohttp

sqlmodel

requests-cache

beautifulsoup4
lxml
faust-cchardet

playwright